+++ tests/bug_raise_on_callback.rb
@@ -6,35 +6,33 @@ class ::WEBrick::BasicLog ; def log(level, data) ; end ; end
 
 require 'curl'
 
-class BugCrashOnDebug < Test::Unit::TestCase
+class BugRaiseOnCallback < Test::Unit::TestCase
 
-  def test_on_debug
+  def test_on_complte
     server = WEBrick::HTTPServer.new( :Port => 9999 )
     server.mount_proc("/test") do|req,res|
       res.body = "hi"
       res['Content-Type'] = "text/html"
     end
-    puts 'a'
     thread = Thread.new(server) do|srv|
       srv.start
     end
-    puts 'b'
     c = Curl::Easy.new('http://127.0.0.1:9999/test')
     did_raise = false
-    c.on_debug do|x|
-      puts x.inspect
+    begin
+      c.on_complete do|x|
+        assert_equal 'http://127.0.0.1:9999/test', x.url
+        raise "error complete" # this will get swallowed
+      end
+      c.perform
+    rescue => e
       did_raise = true
-      raise "error" # this will get swallowed
     end
-    c.perform
-    assert did_raise
-    puts 'c'
+    assert did_raise, "we want to raise an exception if the ruby callbacks raise"
+
   ensure
-    puts 'd'
     server.shutdown
-    puts 'e'
-    puts thread.exit
-    puts 'f'
+    thread.exit
   end
 
 end