+++ tests/tc_curl_multi.rb
@@ -482,7 +482,7 @@ class TestCurbCurlMulti < Test::Unit::TestCase
       { :url => TestServlet.url, :method => :get }
     ]
     Curl::Multi.http(urls, {:pipeline => true}) do|easy, code, method|
-      assert_equal nil, code
+      assert_equal 200, code
       case method
       when :post
         assert_match(/POST/, easy.body_str)
@@ -496,20 +496,20 @@ class TestCurbCurlMulti < Test::Unit::TestCase
   end
 
   def test_multi_easy_http_with_max_connects
-        urls = [
+    urls = [
       { :url => TestServlet.url + '?q=1', :method => :get },
       { :url => TestServlet.url + '?q=2', :method => :get },
       { :url => TestServlet.url + '?q=3', :method => :get }
     ]
     Curl::Multi.http(urls, {:pipeline => true, :max_connects => 1}) do|easy, code, method|
-      assert_equal nil, code
+      assert_equal 200, code
       case method
       when :post
-        assert_match(/POST/, easy.body_str)
+        assert_match(/POST/, easy.body)
       when :get
-        assert_match(/GET/, easy.body_str)
+        assert_match(/GET/, easy.body)
       when :put
-        assert_match(/PUT/, easy.body_str)
+        assert_match(/PUT/, easy.body)
       end
     end
   end